Fix warnings for 32-bit compiler and misc fixes. (#3896)

+++ pjnath/src/pjnath/ice_session.c
@@ -1677,8 +1677,8 @@ static pj_bool_t on_check_complete(pj_ice_sess *ice,
             }
         }
 
-        LOG5((ice->obj_name, "Check %ld is successful%s",
-             GET_CHECK_ID(&ice->clist, check),
+        LOG5((ice->obj_name, "Check %d is successful%s",
+             (int)GET_CHECK_ID(&ice->clist, check),
              (check->nominated ? " and nominated" : "")));
 
         /* On the first valid pair, we call the callback, if present */
